created foreign key constriants on migration tables

+++ database/migrations/2022_12_21_120115_create_answers_table.php
@@ -15,9 +15,9 @@ return new class extends Migration
   {
     Schema::create('answers', function (Blueprint $table) {
       $table->id();
-      $table->integer('answered_by')->comment('id of the user for which the answer belongs to');
-      $table->integer('survey_id')->comment('id of the survey for which the question being answered belongs to');
-      $table->integer('question_id')->comment('id of the question for which the answer belongs to');
+      $table->unsignedBigInteger('answered_by')->comment('id of the user for which the answer belongs to');
+      $table->unsignedBigInteger('survey_id')->comment('id of the survey for which the question being answered belongs to');
+      $table->unsignedBigInteger('question_id')->comment('id of the question for which the answer belongs to');
       $table->text('answer')->comment('answer to question');
       $table->timestamp('created_at')->useCurrent();
       $table->timestamp('updated_at')->nullable()->useCurrentOnUpdate();
